Reorganize flash gridding and plotting routines
