+++ vis/small_multiples.py
